Add groups support

+++ blue-backend/src/main/java/com/example/demo/repo/UserRepository.java
@@ -5,7 +5,7 @@ import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import com.example.demo.models.User;
+import com.example.demo.entities.User;
 
 /**
  * User repository that will get information about our users.
@@ -14,7 +14,7 @@ import com.example.demo.models.User;
  *
  */
 @Repository
-public interface UserRepository extends JpaRepository<User, Integer> {
+public interface UserRepository extends JpaRepository<User, Long> {
 
 	Optional<User> findByEmail(String email);
 }
